Lazy loading of Schematic Setup panels.
(cherry picked from commit ddc6ecf7be29e8e23bdb7475f1a90373b12072fd)
(cherry picked from commit 6cbc10db813b90292e4a29d1f7c8239c8a1c6765)

+++ common/widgets/wx_treebook.cpp
@@ -90,7 +90,7 @@ bool WX_TREEBOOK::AddLazySubPage( std::function<wxWindow*( wxWindow* aParent )> 
 }
 
 
-wxWindow* WX_TREEBOOK::ResolvePage( int aPage )
+wxWindow* WX_TREEBOOK::ResolvePage( size_t aPage )
 {
     wxWindow* page = GetPage( aPage );
